project-{topic}-plus/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9 # For SQLAlchemy to connect to PostgreSQL
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4 # For password hashing
python-jose[cryptography]==3.3.0 # For JWT
pydantic==2.7.1
pydantic-settings==2.2.1
tenacity==8.4.1 # For database connection retries
loguru==0.7.2 # Optional, for easier logging; otherwise, use built-in logging
fastapi-limiter==0.1.6 # For rate limiting
redis==5.0.4 # For fastapi-limiter and potential caching
httpx==0.27.0 # For integration tests
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0 # For code coverage reports
faker==24.0.0 # For generating seed data and test data
typing_extensions==4.12.0 # Required by Pydantic
async-lru==1.0.4 # For async caching decorator
```